APIRest/requirements.txt
-i https://pypi.org/simple
aniso8601==6.0.0
boto3==1.9.127
botocore==1.12.127
click==7.0
docutils==0.14
flask-restful==0.3.7
flask==1.0.2
itsdangerous==1.1.0
jinja2==2.10
jmespath==0.9.4
markupsafe==1.1.1
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.10.1
pytz==2018.9
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
werkzeug==0.15.2
